Fix tests make tooltip wrap forecast grid
