+++ web/src/features/moreCast2/components/ForecastCell.tsx
@@ -4,6 +4,8 @@ import RemoveCircleIcon from '@mui/icons-material/RemoveCircle'
 import AddBoxIcon from '@mui/icons-material/AddBox'
 import { MEDIUM_GREY } from 'app/theme'
 import ValidatedForecastCell from '@/features/moreCast2/components/ValidatedForecastCell'
+import { useState } from 'react'
+import InvalidCellToolTip from '@/features/moreCast2/components/InvalidCellToolTip'
 
 interface ForecastCellProps {
   disabled: boolean
@@ -19,32 +21,52 @@ const ForecastCell = ({ disabled, label, showGreaterThan, showLessThan, value, v
   if (showGreaterThan && showLessThan) {
     throw Error('ForecastCell cannot show both greater than and less than icons at the same time.')
   }
+
+  const [isHovered, setIsHovered] = useState(false)
+
+  const handleMouseEnter = () => {
+    setIsHovered(true)
+  }
+
+  const handleMouseLeave = () => {
+    setIsHovered(false)
+  }
+
+  const error = validator ? validator(value as string) : ''
+
   return (
-    <Grid container sx={{ justifyContent: 'center', alignItems: 'center' }}>
-      <Grid item xs={2}>
-        {showLessThan && (
-          <Tooltip placement="bottom-end" title="Lower than actual">
-            <RemoveCircleIcon
-              data-testid="forecast-cell-less-than-icon"
-              sx={{ color: MEDIUM_GREY, fontSize: '1.15rem' }}
-            />
-          </Tooltip>
-        )}
+    <InvalidCellToolTip error={error} hoverOnly={(value as string) === ''} hovered={isHovered}>
+      <Grid
+        container
+        sx={{ justifyContent: 'center', alignItems: 'center' }}
+        onMouseEnter={handleMouseEnter}
+        onMouseLeave={handleMouseLeave}
+      >
+        <Grid item xs={2}>
+          {showLessThan && (
+            <Tooltip placement="bottom-end" title="Lower than actual">
+              <RemoveCircleIcon
+                data-testid="forecast-cell-less-than-icon"
+                sx={{ color: MEDIUM_GREY, fontSize: '1.15rem' }}
+              />
+            </Tooltip>
+          )}
+        </Grid>
+        <Grid item xs={8}>
+          <ValidatedForecastCell disabled={disabled} label={label} value={value} validator={validator} />
+        </Grid>
+        <Grid item xs={2} sx={{ marginLeft: 'auto' }}>
+          {showGreaterThan && (
+            <Tooltip placement="bottom-start" title="Higher than actual">
+              <AddBoxIcon
+                data-testid="forecast-cell-greater-than-icon"
+                sx={{ color: MEDIUM_GREY, fontSize: '1.25rem', marginLeft: '2px' }}
+              />
+            </Tooltip>
+          )}
+        </Grid>
       </Grid>
-      <Grid item xs={8}>
-        <ValidatedForecastCell disabled={disabled} label={label} value={value} validator={validator} />
-      </Grid>
-      <Grid item xs={2} sx={{ marginLeft: 'auto' }}>
-        {showGreaterThan && (
-          <Tooltip placement="bottom-start" title="Higher than actual">
-            <AddBoxIcon
-              data-testid="forecast-cell-greater-than-icon"
-              sx={{ color: MEDIUM_GREY, fontSize: '1.25rem', marginLeft: '2px' }}
-            />
-          </Tooltip>
-        )}
-      </Grid>
-    </Grid>
+    </InvalidCellToolTip>
   )
 }
