+++ web/src/features/moreCast2/components/ValidatedForecastCell.tsx
@@ -2,7 +2,6 @@ import React from 'react'
 import { TextField } from '@mui/material'
 import { GridRenderCellParams } from '@mui/x-data-grid-pro'
 import { theme } from 'app/theme'
-import InvalidCellToolTip from '@/features/moreCast2/components/InvalidCellToolTip'
 
 interface ValidatedForecastCellProps {
   disabled: boolean
@@ -14,39 +13,37 @@ interface ValidatedForecastCellProps {
 const ValidatedForecastCell = ({ disabled, label, value, validator }: ValidatedForecastCellProps) => {
   const error = validator ? validator(value as string) : ''
   return (
-    <InvalidCellToolTip error={error} hoverOnly={(value as string) === ''}>
-      <TextField
-        data-testid="validated-forecast-cell"
-        disabled={disabled}
-        size="small"
-        label={label}
-        InputLabelProps={{
-          shrink: true
-        }}
-        sx={{
-          '& .MuiOutlinedInput-root': {
-            backgroundColor: `${theme.palette.common.white}`,
-            '& fieldset': {
-              borderColor: error ? theme.palette.error.main : '#737373',
-              borderWidth: '2px'
-            },
-            '&:hover fieldset': {
-              borderColor: error ? theme.palette.error.main : '#737373'
-            },
-            '&.Mui-focused fieldset': {
-              borderColor: error ? theme.palette.error.main : '#737373',
-              borderWidth: '2px'
-            }
+    <TextField
+      data-testid="validated-forecast-cell"
+      disabled={disabled}
+      size="small"
+      label={label}
+      InputLabelProps={{
+        shrink: true
+      }}
+      sx={{
+        '& .MuiOutlinedInput-root': {
+          backgroundColor: `${theme.palette.common.white}`,
+          '& fieldset': {
+            borderColor: error ? theme.palette.error.main : '#737373',
+            borderWidth: '2px'
           },
-          '& .Mui-disabled': {
-            '& fieldset': {
-              borderWidth: '1px'
-            }
+          '&:hover fieldset': {
+            borderColor: error ? theme.palette.error.main : '#737373'
+          },
+          '&.Mui-focused fieldset': {
+            borderColor: error ? theme.palette.error.main : '#737373',
+            borderWidth: '2px'
           }
-        }}
-        value={value}
-      ></TextField>
-    </InvalidCellToolTip>
+        },
+        '& .Mui-disabled': {
+          '& fieldset': {
+            borderWidth: '1px'
+          }
+        }
+      }}
+      value={value}
+    ></TextField>
   )
 }
