+++ web/src/features/moreCast2/components/editInputCell.test.tsx
@@ -1,4 +1,4 @@
-import { render, fireEvent, within } from '@testing-library/react'
+import { render, fireEvent, within, act } from '@testing-library/react'
 import { useDispatch } from 'react-redux'
 import { GridApiContext, GridCellMode, GridTreeNodeWithRender } from '@mui/x-data-grid-pro'
 import { EditInputCell } from '@/features/moreCast2/components/EditInputCell'
@@ -63,7 +63,7 @@ describe('EditInputCell', () => {
     expect(input).toHaveFocus()
   })
 
-  test('should call setEditCellValue on value change', () => {
+  test('should call setEditCellValue on value change', async () => {
     const { getByTestId } = render(
       <GridApiContext.Provider value={apiMock}>
         <EditInputCell {...defaultProps} id={1} value="10" field="test" hasFocus={false} error="" />
@@ -72,12 +72,13 @@ describe('EditInputCell', () => {
 
     const input = within(getByTestId('forecast-edit-cell')).getByRole('spinbutton') as HTMLInputElement
     expect(input.value).toBe('10')
-    // Change the value and fire the event
-    fireEvent.change(input, { target: { value: '20' } })
+    await act(async () => {
+      fireEvent.change(input, { target: { value: '20' } })
+    })
     expect(mockSetEditCellValue).toHaveBeenCalledWith({ id: 1, field: 'test', value: '20' })
   })
 
-  test('should call stopCellEditMode on blur', () => {
+  test('should call stopCellEditMode on blur', async () => {
     const { getByTestId } = render(
       <GridApiContext.Provider value={apiMock}>
         <EditInputCell {...defaultProps} id={1} value="10" field="test" hasFocus={false} error="" />
@@ -86,12 +87,15 @@ describe('EditInputCell', () => {
 
     const input = within(getByTestId('forecast-edit-cell')).getByRole('spinbutton') as HTMLInputElement
     input.focus()
-    fireEvent.blur(input)
+
+    await act(async () => {
+      fireEvent.blur(input)
+    })
 
     expect(mockStopCellEditMode).toHaveBeenCalledWith({ id: 1, field: 'test' })
   })
 
-  test('should handle Escape key press', () => {
+  test('should handle Escape key press', async () => {
     const { getByTestId } = render(
       <GridApiContext.Provider value={apiMock}>
         <EditInputCell {...defaultProps} id={1} value="10" field="test" hasFocus={false} error="" />
@@ -100,12 +104,15 @@ describe('EditInputCell', () => {
 
     const input = within(getByTestId('forecast-edit-cell')).getByRole('spinbutton') as HTMLInputElement
     input.focus()
-    fireEvent.keyDown(input, { key: 'Escape', code: 'Escape', charCode: 27 })
+
+    await act(async () => {
+      fireEvent.keyDown(input, { key: 'Escape', code: 'Escape', charCode: 27 })
+    })
 
     expect(mockStopCellEditMode).toHaveBeenCalledWith({ id: 1, field: 'test' })
   })
 
-  test('should not call stopCellEditMode when Escape key is pressed and there is an error', () => {
+  test('should not call stopCellEditMode when Escape key is pressed and there is an error', async () => {
     const { getByTestId } = render(
       <GridApiContext.Provider value={{ current: apiMock }}>
         <EditInputCell {...defaultProps} error="Test error" />
@@ -114,8 +121,10 @@ describe('EditInputCell', () => {
 
     const input = within(getByTestId('forecast-edit-cell')).getByRole('spinbutton') as HTMLInputElement
     input.focus()
-    // Simulate Escape key press
-    fireEvent.keyDown(input, { key: 'Escape', code: 'Escape', charCode: 27 })
+
+    await act(async () => {
+      fireEvent.keyDown(input, { key: 'Escape', code: 'Escape', charCode: 27 })
+    })
 
     // Verify that stopCellEditMode was not called
     expect(mockStopCellEditMode).not.toHaveBeenCalled()
